Fix thread specific data in nscd(8) (and possibly ldap_cachemgr(8))
`nscd` and `ldap_cachemgr` are both normally 32bit, and made assumptions that
`int` and a pointer are the same size, to store and retrieve their
thread-specific data.  Explicitly used `intptr_t` so that this is true
regardless of ISA.

I have no LDAP environment and have not tested `ldap_cachemgr`.

+++ usr/src/cmd/ldapcachemgr/cachemgr.c
@@ -188,8 +188,8 @@ get_clearance(int callnumber)
 int
 release_clearance(int callnumber)
 {
-	int	which;
-	sema_t	*table_sema = NULL;
+	intptr_t	which;
+	sema_t		*table_sema = NULL;
 
 	(void) thr_getspecific(lookup_state_key, (void**)&which);
 	if (which == 0) /* from common pool */ {
